+++ demos/example-vite/CHANGELOG.md
@@ -1,5 +1,11 @@
 # example-vite
 
+## 0.0.11
+
+### Patch Changes
+
+- @journeyapps/powersync-sdk-web@0.3.3
+
 ## 0.0.10
 
 ### Patch Changes