+++ packages/powersync-attachments/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @journeyapps/powersync-attachments
 
+## 2.0.0
+
+### Major Changes
+
+- d223454: Renamed abstract `attachmentIds` to `onAttachmentIdsChange` in `AbstractAttachmentQueue`, changed to use callback signature instead of an AsyncGenerator.
+
+### Patch Changes
+
+- Updated dependencies [fd7ebc8]
+  - @journeyapps/powersync-sdk-common@1.4.0
+
 ## 1.1.3
 
 ### Patch Changes